Move team database stuff to new 'db' subpackage
