+++ app/src/main/java/rec/games/pokemon/teambuilder/db/AppDatabase.java
@@ -1,4 +1,4 @@
-package rec.games.pokemon.teambuilder;
+package rec.games.pokemon.teambuilder.db;
 
 import android.arch.persistence.room.Database;
 import android.arch.persistence.room.Room;
